chore: split env tests
